Add TransferAngle
- Move MicroUtility to Utilities folder
- Move AeroForces to a new file -> Utilities\AeroForces.cs
- Add new class TransferInfo -> Utilities\TransferInfo.cs
- Add TransferAngle
